data_layer/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.34.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keybert==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.54.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
